Added authorization code to other operations

+++ src/main/java/com/skb/course/apis/libraryapis/author/AuthorController.java
@@ -5,6 +5,7 @@ import com.skb.course.apis.libraryapis.exception.LibraryResourceBadRequestExcept
 import com.skb.course.apis.libraryapis.exception.LibraryResourceNotFoundException;
 import com.skb.course.apis.libraryapis.author.Author;
 import com.skb.course.apis.libraryapis.author.AuthorService;
+import com.skb.course.apis.libraryapis.exception.LibraryResourceUnauthorizedException;
 import com.skb.course.apis.libraryapis.util.LibraryApiUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,20 +36,25 @@ public class AuthorController {
         if(!LibraryApiUtils.doesStringValueExist(traceId)) {
             traceId = UUID.randomUUID().toString();
         }
-
+        logger.debug("Added TraceId: {}", traceId);
         return new ResponseEntity<>(authorService.getAuthor(authorId, traceId), HttpStatus.OK);
     }
 
     @PostMapping
     public ResponseEntity<?> addAuthor(@Valid @RequestBody Author author,
-                                          @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId)
-            throws LibraryResourceAlreadyExistException {
+                                          @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId,
+                                          @RequestHeader(value = "Authorization") String bearerToken)
+            throws LibraryResourceAlreadyExistException, LibraryResourceUnauthorizedException {
 
         logger.debug("Request to add Author: {}", author);
         if(!LibraryApiUtils.doesStringValueExist(traceId)) {
             traceId = UUID.randomUUID().toString();
         }
         logger.debug("Added TraceId: {}", traceId);
+        if(!LibraryApiUtils.isUserAdmin(bearerToken)) {
+            logger.error(LibraryApiUtils.getUserIdFromClaim(bearerToken) + " attempted to add a Author. Disallowed because user is not Admin");
+            throw new LibraryResourceUnauthorizedException(traceId, "User not allowed to Add a Author");
+        }
         authorService.addAuthor(author, traceId);
 
         logger.debug("Returning response for TraceId: {}", traceId);
@@ -57,47 +63,59 @@ public class AuthorController {
 
     @PutMapping(path = "/{authorId}")
     public ResponseEntity<?> updateAuthor(@PathVariable Integer authorId,
-                                             @Valid @RequestBody Author author,
-                                             @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId)
-            throws LibraryResourceNotFoundException {
+                                          @Valid @RequestBody Author author,
+                                          @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId,
+                                          @RequestHeader(value = "Authorization") String bearerToken)
+            throws LibraryResourceNotFoundException, LibraryResourceUnauthorizedException {
 
         if(!LibraryApiUtils.doesStringValueExist(traceId)) {
             traceId = UUID.randomUUID().toString();
         }
+        logger.debug("Added TraceId: {}", traceId);
+        if(!LibraryApiUtils.isUserAdmin(bearerToken)) {
+            logger.error(LibraryApiUtils.getUserIdFromClaim(bearerToken) + " attempted to update a Author. Disallowed because user is not Admin");
+            throw new LibraryResourceUnauthorizedException(traceId, "User not allowed to Add a Author");
+        }
 
         author.setAuthorId(authorId);
         authorService.updateAuthor(author, traceId);
-
+        logger.debug("Returning response for TraceId: {}", traceId);
         return new ResponseEntity<>(author, HttpStatus.OK);
     }
 
     @DeleteMapping(path = "/{authorId}")
     public ResponseEntity<?> deleteAuthor(@PathVariable Integer authorId,
-                                             @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId)
-            throws LibraryResourceNotFoundException {
+                                          @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId,
+                                          @RequestHeader(value = "Authorization") String bearerToken)
+            throws LibraryResourceNotFoundException, LibraryResourceUnauthorizedException {
 
         if(!LibraryApiUtils.doesStringValueExist(traceId)) {
             traceId = UUID.randomUUID().toString();
         }
-
+        if(!LibraryApiUtils.isUserAdmin(bearerToken)) {
+            logger.error(LibraryApiUtils.getUserIdFromClaim(bearerToken) + " attempted to delete a Author. Disallowed because user is not Admin");
+            throw new LibraryResourceUnauthorizedException(traceId, "User not allowed to Add a Author");
+        }
+        logger.debug("Added TraceId: {}", traceId);
         authorService.deleteAuthor(authorId, traceId);
+        logger.debug("Returning response for TraceId: {}", traceId);
         return new ResponseEntity<>(HttpStatus.ACCEPTED);
     }
 
     @GetMapping(path = "/search")
     public ResponseEntity<?> searchAuthor(@RequestParam String firstName, @RequestParam String lastName,
-                                             @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId)
+                                          @RequestHeader(value = "Trace-Id", defaultValue = "") String traceId)
             throws LibraryResourceBadRequestException {
 
         if(!LibraryApiUtils.doesStringValueExist(traceId)) {
             traceId = UUID.randomUUID().toString();
         }
-
+        logger.debug("Added TraceId: {}", traceId);
         if(!LibraryApiUtils.doesStringValueExist(firstName) && !LibraryApiUtils.doesStringValueExist(lastName)) {
             logger.error("TraceId: {}, Please enter at least one search criteria to search Authors!!", traceId);
             throw new LibraryResourceBadRequestException(traceId, "Please enter a name to search Author.");
         }
-
+        logger.debug("Returning response for TraceId: {}", traceId);
         return new ResponseEntity<>(authorService.searchAuthor(firstName, lastName, traceId), HttpStatus.OK);
     }
 }
